--- .github/workflows/ci-polars.yml
@@ -1,12 +1,0 @@
-name: Continuous Integration (Master)
-
-on:
-  push:
-    branches:
-      - "polars2"
-
-jobs:
-  merge_gate:
-    name: Checks
-    uses: ./.github/workflows/merge-gate.yml
-    secrets: inherit
