+++ doc/changes/unreleased.md
@@ -1,1 +1,3 @@
 # Unreleased
+
+- [#214](https://github.com/exasol/pyexasol/issues/214): Added support to EXPORT and IMPORT from Exasol to Polars DataFrames
