+++ doc/user_guide/dependencies.rst
@@ -12,7 +12,8 @@ Core dependencies
 Optional dependencies
 +++++++++++++++++++++
 
-- ``pandas`` is required for :ref:`http_transport` functions working with data frames
+- ``pandas`` is required for :ref:`http_transport` functions working with pandas data frames
+- ``polars`` is required for :ref:`http_transport` functions working with polars data frames
 - ``ujson`` is required for ``json_lib=ujson`` to improve JSON parsing performance
 - ``rapidjson`` is required for ``json_lib=rapidjson`` to improve JSON parsing performance
 - ``orjson`` is required for ``json_lib=orjson`` to improve JSON parsing performance
